+++ core/model/application.go
@@ -12,10 +12,10 @@ const (
 	TypeApplication logutils.MessageDataType = "application"
 	//TypePermission ...
 	TypePermission logutils.MessageDataType = "permission"
-	//TypeApplicationRole ...
-	TypeApplicationRole logutils.MessageDataType = "application role"
-	//TypeApplicationGroup ...
-	TypeApplicationGroup logutils.MessageDataType = "application group"
+	//TypeAppOrgRole ...
+	TypeAppOrgRole logutils.MessageDataType = "application organization role"
+	//TypeAppOrgGroup ...
+	TypeAppOrgGroup logutils.MessageDataType = "application organization group"
 	//TypeOrganization ...
 	TypeOrganization logutils.MessageDataType = "organization"
 	//TypeApplicationOrganization ...
@@ -31,62 +31,55 @@ type Permission struct {
 	ID   string `bson:"_id"`
 	Name string `bson:"name"`
 
-	ServiceIDs []string `bson:"service_ids"`
+	ServiceID string   `bson:"service_id"`
+	Assigners []string `bson:"assigners"`
 
 	DateCreated time.Time  `bson:"date_created"`
 	DateUpdated *time.Time `bson:"date_updated"`
 }
 
-func (c Permission) getServiceIDs() string {
-	s := ""
-	if c.ServiceIDs == nil || len(c.ServiceIDs) == 0 {
-		return s
-	}
-	for _, id := range c.ServiceIDs {
-		s += id + ","
-	}
-
-	return s
-}
-
 func (c Permission) String() string {
-	return fmt.Sprintf("[ID:%s\nName:%s\nServiceIDs:%s]", c.ID, c.Name, c.getServiceIDs())
+	return fmt.Sprintf("[ID:%s\nName:%s\nServiceID:%s]", c.ID, c.Name, c.ServiceID)
 }
 
-//ApplicationRole represents application role entity. It is a collection of permissions
-type ApplicationRole struct {
+//AppOrgRole represents application organization role entity. It is a collection of permissions
+type AppOrgRole struct {
 	ID          string
 	Name        string
 	Description string
 
+	System bool
+
 	Permissions []Permission
 
-	Application Application
+	AppOrg ApplicationOrganization
 
 	DateCreated time.Time
 	DateUpdated *time.Time
 }
 
-func (c ApplicationRole) String() string {
-	return fmt.Sprintf("[ID:%s\tName:%s\tPermissions:%s\tApplication:%s]", c.ID, c.Name, c.Permissions, c.Application.Name)
+func (c AppOrgRole) String() string {
+	return fmt.Sprintf("[ID:%s\tName:%s\tPermissions:%s\tAppOrg:%s]", c.ID, c.Name, c.Permissions, c.AppOrg.ID)
 }
 
-//ApplicationGroup represents application group entity. It is a collection of users
-type ApplicationGroup struct {
+//AppOrgGroup represents application organization group entity. It is a collection of users
+type AppOrgGroup struct {
 	ID   string
 	Name string
 
-	Permissions []Permission
-	Roles       []ApplicationRole
+	System bool
 
-	Application Application
+	Permissions []Permission
+	Roles       []AppOrgRole
+
+	AppOrg ApplicationOrganization
 
 	DateCreated time.Time
 	DateUpdated *time.Time
 }
 
-func (cg ApplicationGroup) String() string {
-	return fmt.Sprintf("[ID:%s\nName:%s\nApplication:%s]", cg.ID, cg.Name, cg.Application.Name)
+func (cg AppOrgGroup) String() string {
+	return fmt.Sprintf("[ID:%s\nName:%s\nAppOrg:%s]", cg.ID, cg.Name, cg.AppOrg.ID)
 }
 
 //Application represents users application entity - safer community, uuic, etc
@@ -141,6 +134,8 @@ type ApplicationOrganization struct {
 
 	Application  Application
 	Organization Organization
+
+	ServicesIDs []string //which services are used for this app/org
 
 	IdentityProvidersSettings []IdentityProviderSetting