+++ core/model/user.go
@@ -34,8 +34,7 @@ const (
 type Account struct {
 	ID string //this is ID for the account
 
-	Application  Application
-	Organization Organization
+	AppOrg ApplicationOrganization
 
 	Permissions []Permission
 	Roles       []AccountRole