+++ CHANGELOG.md
@@ -8,6 +8,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Handle groups mappings from OIDC integration [#276](https://github.com/rokwire/core-building-block/issues/276)
+- Add refresh token abuse detection to login sessions [#257](https://github.com/rokwire/core-building-block/issues/257)
+
+### Changed
+- Expose transaction interface on storage adapter [#285](https://github.com/rokwire/core-building-block/issues/285)
+
+### Added
+- Define unified responses for the APIs [#286](https://github.com/rokwire/core-building-block/issues/286)
+
+## [1.11.0] - 2021-11-04
+### Added
+- Set up permission groups [#25](https://github.com/rokwire/core-building-block/issues/25)
+
+## [1.10.0] - 2021-11-03
+### Added
+- Merge client and Profile BB profiles and preferences [#228](https://github.com/rokwire/core-building-block/issues/228)
 
 ## [1.9.0] - 2021-11-01
 ### Added
@@ -35,7 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.5.0] - 2021-10-15
 ### Fixed
-- Permission authorization failing on all endpoints in Docker[#239](https://github.com/rokwire/core-building-block/issues/239)
+- Permission authorization failing on all endpoints in Docker [#239](https://github.com/rokwire/core-building-block/issues/239)
 
 ### Changed
 - Switch to ROKWIRE open source libraries [#232](https://github.com/rokwire/core-building-block/issues/232)
@@ -128,7 +143,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.9.0...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.11.0...HEAD
+[1.11.0]: https://github.com/rokwire/core-building-block/compare/v1.10.0...v1.11.0
+[1.10.0]: https://github.com/rokwire/core-building-block/compare/v1.9.0...v1.10.0
 [1.9.0]: https://github.com/rokwire/core-building-block/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/rokwire/core-building-block/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/rokwire/core-building-block/compare/v1.6.0...v1.7.0